requirements.txt

numpy
tensorflow
torch
torchvision
transformers
pandas
litellm
torch-geometric
dgl
dgllife
pytdc
keras
xgboost
anthropic